split out API manager access layer, finish getKey
