+++ smrt-server-analysis/src/main/scala/com/pacbio/secondary/smrtserver/tools/ApiManagerAccessLayer.scala
@@ -1,0 +1,175 @@
+package com.pacbio.secondary.smrtserver.tools
+
+import java.nio.file.Path
+import java.security.SecureRandom
+import java.security.cert.X509Certificate
+import javax.net.ssl.{SSLContext, X509TrustManager, TrustManager}
+
+import scala.concurrent.duration._
+import scala.concurrent.Future
+import scala.util.Random
+
+import akka.actor.ActorSystem
+import akka.io.IO
+import akka.pattern.ask
+import akka.util.Timeout
+
+import spray.http._
+import spray.can.Http
+import spray.client.pipelining._
+import spray.httpx.marshalling.BasicMarshallers._
+import spray.httpx.SprayJsonSupport
+
+import com.typesafe.scalalogging.LazyLogging
+
+import org.wso2.carbon.apimgt.rest.api.store
+import org.wso2.carbon.apimgt.rest.api.publisher
+
+
+class ApiManagerAccessLayer(host: String, portOffset: Int = 0, user: String, pass: String)(implicit actorSystem: ActorSystem) extends LazyLogging {
+
+  import actorSystem.dispatcher
+
+  import ApiManagerJsonProtocols._
+  import SprayJsonSupport._
+
+  implicit val timeout: Timeout = 30.seconds
+
+  val ADMIN_PORT = 9443
+  val API_PORT = 8243
+
+
+  implicit val sslContext = {
+    // Create a trust manager that does not validate certificate chains.
+    val permissiveTrustManager: TrustManager = new X509TrustManager() {
+      override def checkClientTrusted(chain: Array[X509Certificate], authType: String): Unit = {
+      }
+      override def checkServerTrusted(chain: Array[X509Certificate], authType: String): Unit = {
+      }
+      override def getAcceptedIssuers(): Array[X509Certificate] = {
+        null
+      }
+    }
+
+    val initTrustManagers = Array(permissiveTrustManager)
+    val ctx = SSLContext.getInstance("TLS")
+    ctx.init(null, initTrustManagers, new SecureRandom())
+    ctx
+  }
+
+  // request pipeline for the admin port
+  val adminPipe: Future[SendReceive] =
+    for (
+      Http.HostConnectorInfo(connector, _) <-
+        IO(Http) ? Http.HostConnectorSetup(host, port = ADMIN_PORT + portOffset, sslEncryption = true)
+    ) yield sendReceive(connector)
+
+  // request pipeline for the API port
+  val apiPipe: Future[SendReceive] =
+    for (
+      Http.HostConnectorInfo(connector, _) <-
+        IO(Http) ? Http.HostConnectorSetup(host, port = API_PORT + portOffset, sslEncryption = true)
+    ) yield sendReceive(connector)
+
+  def register(): Future[ClientRegistrationResponse] = {
+    val body = ClientRegistrationRequest("foo", Random.alphanumeric.take(20).mkString, "Production", user, "password refresh_token", true)
+
+    val request = (
+      Post(s"/client-registration/v0.10/register", body)
+        ~> addCredentials(BasicHttpCredentials(user, pass))
+    )
+
+    adminPipe.flatMap(_(request)).map(unmarshal[ClientRegistrationResponse])
+  }
+
+  // the consumer key and secret can come from the dynamic client
+  // registration mechanism (the register() method) or they can come
+  // from some existing configuration, like the UI's app-config.json
+  def login(consumerKey: String, consumerSecret: String, scopes: String): Future[OauthToken] = {
+    val body = FormData(Map(
+      "grant_type" -> "password",
+      "username" -> user,
+      "password" -> pass,
+      "scope" -> scopes
+    ))
+
+    val request = (
+      Post(s"/token", body)
+        ~> addCredentials(BasicHttpCredentials(consumerKey, consumerSecret))
+    )
+    apiPipe.flatMap(_(request)).map(unmarshal[OauthToken])
+  }
+
+  // Store APIs
+  def putApplication(app: store.models.Application, token: OauthToken): Future[HttpResponse] = {
+    val request = (
+      Put(s"/api/am/store/v0.10/applications/${app.applicationId.get}", app)
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request))
+  }
+
+  def getApplication(applicationId: String, token: OauthToken): Future[store.models.Application] = {
+    val request = (
+      Get(s"/api/am/store/v0.10/applications/${applicationId}")
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request)).map(unmarshal[store.models.Application])
+  }
+
+  def searchApplications(name: String, token: OauthToken): Future[store.models.ApplicationList] = {
+    val request = (
+      Get(s"/api/am/store/v0.10/applications?query=${name}")
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request)).map(unmarshal[store.models.ApplicationList])
+  }
+
+
+  // publisher APIs
+  def searchApis(name: String, token: OauthToken): Future[publisher.models.APIList] = {
+    val request = (
+      Get(s"/api/am/publisher/v0.10/apis?query=${name}")
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request)).map(unmarshal[publisher.models.APIList])
+  }
+
+  def getApiDetails(id: String, token: OauthToken): Future[publisher.models.API] = {
+    val request = (
+      Get(s"/api/am/publisher/v0.10/apis/${id}")
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request)).map(unmarshal[publisher.models.API])
+  }
+
+  def putApiDetails(api: publisher.models.API, token: OauthToken): Future[publisher.models.API] = {
+    val request = (
+      Put(s"/api/am/publisher/v0.10/apis/${api.id.get}", api)
+        ~> addHeader("Authorization", s"Bearer ${token.access_token}")
+    )
+    adminPipe.flatMap(_(request)).map(unmarshal[publisher.models.API])
+  }
+
+  def updateEndpoints(api: publisher.models.APIInfo, prodEndpoint: String, devEndpoint: String, token: OauthToken): Future[publisher.models.API] = {
+    val newEndpointConfig = s"""
+{
+  "production_endpoints": {
+    "url": "${prodEndpoint}",
+    "config": null
+  },
+  "sandbox_endpoints\":{
+    "url": "${devEndpoint}",
+    "config": null
+  },
+  "endpoint_type": "http"
+}
+"""
+
+    getApiDetails(api.id.get, token).flatMap({ details =>
+      val updated = details.copy(endpointConfig = newEndpointConfig)
+      putApiDetails(updated, token)
+    })
+  }
+
+}
